requirements.txt
claude-code-sdk>=0.0.20
anthropic>=0.34.0
python-dotenv>=1.0.0
asyncio
requests>=2.31.0
beautifulsoup4>=4.12.0
arxiv>=2.1.0
scholarly>=1.7.11
PyPDF2>=3.0.1
aiohttp>=3.8.0
slack-sdk>=3.23.0
pyzotero>=1.5.0